Added new method, view(), for use when debugging
I have added in a new method that can be called in the same way that `stream` is in order to print raw HTML to the browser.  
This is useful for debugging the HTML to view what the code will look like - roughly.

Also, I have modified the comments on the config array so that they reflect the wording from the Mpdf documentation.